Divide TestPsci into several modules
- TestPsci.TestEnv: About init'ing and running PSCi test env
- TestPsci.CommandTest: About PSCi command tests
- TestPsci.CompletionTest: About PSCi completion tests
